Remove necessity for the JsBridge to know about the WebView
